requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pytz==2024.1
Jinja2==3.1.4
openai==1.40.2
